feat: 상세보기뷰 바텀시트프래그먼트 구현
